家人列表

+++ src/main/java/com/zhang/genealogy/model/Personnel.java
@@ -54,7 +54,7 @@ public class Personnel {
     /**
      * 配偶ID
      */
-    private Integer coupleId;
+    private Long coupleId;
 
     /**
      * 世代
